tools/test/vagrant: add --install_version=X option

+++ tools/test/vagrant/resources/shell_provisioners/install_ubuntu_packages.sh
@@ -4,13 +4,24 @@
 test -d /var/cfengine && exit 0
 
 cd /tmp
-wget -q http://cfengine.com/pub/gpg.key
-apt-key add gpg.key
-rm gpg.key
 
-apt-get -qq update
-apt-get -qq -y install python-software-properties
-apt-get -qq -y install software-properties-common
-add-apt-repository http://cfengine.com/pub/apt
-apt-get -qq update
-apt-get -qq -y install cfengine-community
+if [ -z "$1" ]; then
+    echo Installing Community packages
+
+    wget -q http://cfengine.com/pub/gpg.key
+    apt-key add gpg.key
+    rm gpg.key
+
+    apt-get -qq update
+    apt-get -qq -y install python-software-properties
+    apt-get -qq -y install software-properties-common
+    add-apt-repository http://cfengine.com/pub/apt
+    apt-get -qq update
+    apt-get -qq -y install cfengine-community
+else
+    while [ -n "$1" ]; do
+        echo Installing requested package $1
+        /usr/bin/dpkg -i $1
+        shift
+    done
+fi
